Comment update
The `param targets` fed to classifier is the target label after being fit_transform(). The shape of `param targets` should be [batch size, 2]
For example, we have classification task with 3 categories: 0,1,2.  
The list of labels [0, 1, 2] will be transformed to [array([1, 0]), array([0.5, 0.5]), array([0, 1])]
I am sure whether I misunderstand but `n_classes` here may not referred to number of classes?